Refactor schemas into their own files
